renderItem Function - react native papers - ListStationComponent - MapComponent - src file
