+++ package.json
@@ -11,12 +11,13 @@
   "dependencies": {
     "axios": "^1.2.1",
     "expo": "~47.0.8",
+    "expo-constants": "~14.0.2",
+    "expo-location": "~15.0.1",
     "expo-status-bar": "~1.4.2",
     "react": "18.1.0",
     "react-native": "0.70.5",
-    "expo-location": "~15.0.1",
-    "expo-constants": "~14.0.2",
-    "react-native-maps": "1.3.2"
+    "react-native-maps": "1.3.2",
+    "react-native-paper": "^5.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
